parser_metal_a.py
import requests
from bs4 import BeautifulSoup
from models import UniversalProduct

def parse_source_a():
    """Парсер для поставщика металлопроката А (обезличено)"""
    headers = {
        "Accept": "*/*",
        "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/114.0.0.0 Safari/537.36"
    }
    
    # URL заменен на заглушку
    link = "https://provider-a-catalog.example.com/metalloprokat/"
    
    try:
        # В реальном проекте здесь был бы запрос, для GitHub оставляем структуру
        # res = requests.get(link, headers=headers, timeout=10)
        # soup = BeautifulSoup(res.content, "html.parser")
        
        results = []
        # Имитация логики парсинга для демонстрации структуры
        sample_items = ["Арматура А500С", "Труба профильная 40х40", "Лист горячекатаный"]
        
        for item_name in sample_items:
            product = UniversalProduct(
                name=item_name,
                source="Metal_Provider_A",
                category="Металлопрокат",
                raw_price=None, # Сработает генерация
                stock_status="В наличии"
            )
            results.append(product)
        
        return results
    except Exception as e:
        print(f"Ошибка при парсинге Provider A: {e}")
        return []
